+++ src/main/java/org/kasbench/globeco_trade_service/dto/ExecutionPutDTO.java
@@ -11,7 +11,7 @@ public class ExecutionPutDTO {
     private Integer tradeTypeId;
     private Integer tradeOrderId;
     private Integer destinationId;
-    private Short quantityOrdered;
+    private BigDecimal quantityOrdered;
     private BigDecimal quantityPlaced;
     private BigDecimal quantityFilled;
     private BigDecimal limitPrice;
@@ -60,10 +60,10 @@ public class ExecutionPutDTO {
     public void setDestinationId(Integer destinationId) {
         this.destinationId = destinationId;
     }
-    public Short getQuantityOrdered() {
+    public BigDecimal getQuantityOrdered() {
         return quantityOrdered;
     }
-    public void setQuantityOrdered(Short quantityOrdered) {
+    public void setQuantityOrdered(BigDecimal quantityOrdered) {
         this.quantityOrdered = quantityOrdered;
     }
     public BigDecimal getQuantityPlaced() {
